+++ litebird_sim/mapmaking/destriper.py
@@ -33,6 +33,8 @@ from .common import (
     cholesky,
     solve_cholesky,
     estimate_cond_number,
+    _build_mask_detector_split,
+    _build_mask_time_split,
 )
 
 if MPI_ENABLED:
@@ -326,9 +328,11 @@ class DestriperResult:
     stopping_factor: Optional[float]
     history_of_stopping_factors: Optional[List[float]]
     destriped_map: Optional[npt.ArrayLike]
-    converged: bool
+    converged: Union[bool, str]
     elapsed_time_s: float
     bytes_in_temporary_buffers: int
+    detector_split: Optional[str] = "full"
+    time_split: Optional[str] = "full"
 
 
 def _sum_components_into_obs(
@@ -356,6 +360,8 @@ def _accumulate_nobs_matrix(
     pix_idx: npt.ArrayLike,  # Shape: (Ndet, 1)
     psi_angle_rad: npt.ArrayLike,  # Shape: (Ndet, 1)
     weights: npt.ArrayLike,  # Shape: (N_det,)
+    d_mask: npt.ArrayLike,
+    t_mask: npt.ArrayLike,
     nobs_matrix: npt.ArrayLike,  # Shape: (N_pix, 6)
 ) -> None:
     """
@@ -369,24 +375,32 @@ def _accumulate_nobs_matrix(
 
     assert pix_idx.shape == psi_angle_rad.shape
 
+    assert pix_idx.shape[0] == d_mask.shape[0]
+
     num_of_detectors = pix_idx.shape[0]
 
     for det_idx in range(num_of_detectors):
+        if not d_mask[det_idx]:
+            continue
+
         inv_sigma = 1.0 / np.sqrt(weights[det_idx])
         inv_sigma2 = inv_sigma * inv_sigma
 
         # Fill the lower triangle of M_i only for i = 1…N_pix
-        for cur_pix_idx, cur_psi in zip(pix_idx[det_idx], psi_angle_rad[det_idx]):
-            cos_over_sigma = np.cos(2 * cur_psi) * inv_sigma
-            sin_over_sigma = np.sin(2 * cur_psi) * inv_sigma
-            cur_matrix = nobs_matrix[cur_pix_idx]
+        for cur_pix_idx, cur_psi, cur_t_mask in zip(
+            pix_idx[det_idx], psi_angle_rad[det_idx], t_mask
+        ):
+            if cur_t_mask:
+                cos_over_sigma = np.cos(2 * cur_psi) * inv_sigma
+                sin_over_sigma = np.sin(2 * cur_psi) * inv_sigma
+                cur_matrix = nobs_matrix[cur_pix_idx]
 
-            cur_matrix[0] += inv_sigma2
-            cur_matrix[1] += cos_over_sigma * inv_sigma
-            cur_matrix[2] += cos_over_sigma * cos_over_sigma
-            cur_matrix[3] += sin_over_sigma * inv_sigma
-            cur_matrix[4] += sin_over_sigma * cos_over_sigma
-            cur_matrix[5] += sin_over_sigma * sin_over_sigma
+                cur_matrix[0] += inv_sigma2
+                cur_matrix[1] += cos_over_sigma * inv_sigma
+                cur_matrix[2] += cos_over_sigma * cos_over_sigma
+                cur_matrix[3] += sin_over_sigma * inv_sigma
+                cur_matrix[4] += sin_over_sigma * cos_over_sigma
+                cur_matrix[5] += sin_over_sigma * sin_over_sigma
 
 
 @njit(parallel=True)
@@ -450,18 +464,24 @@ def _build_nobs_matrix(
     obs_list: List[Observation],
     ptg_list: List[npt.ArrayLike],
     psi_list: List[npt.ArrayLike],
+    dm_list: List[npt.ArrayLike],
+    tm_list: List[npt.ArrayLike],
 ) -> NobsMatrix:
     # Instead of a shape like (Npix, 3, 3), i.e., one 3×3 matrix per each
     # pixel, we only store the lower triangular part in a 6-element array.
     # In this way we reduce the memory usage by ~30% and the code is faster too.
     nobs_matrix = np.zeros((hpx.npix(), 6))  # Do not use np.empty() here!
 
-    for cur_obs, cur_ptg, cur_psi in zip(obs_list, ptg_list, psi_list):
+    for cur_obs, cur_ptg, cur_psi, cur_d_mask, cur_t_mask in zip(
+        obs_list, ptg_list, psi_list, dm_list, tm_list
+    ):
         _accumulate_nobs_matrix(
             pix_idx=cur_obs.destriper_pixel_idx,
             psi_angle_rad=cur_obs.destriper_pol_angle_rad,
             weights=cur_obs.destriper_weights,
             nobs_matrix=nobs_matrix,
+            d_mask=cur_d_mask,
+            t_mask=cur_t_mask,
         )
 
     # Now we must accumulate the result of every MPI process
@@ -519,6 +539,8 @@ def _update_sum_map_with_tod(
     pol_angle_rad: npt.ArrayLike,
     pixel_idx: npt.ArrayLike,
     weights: npt.ArrayLike,
+    d_mask: npt.ArrayLike,
+    t_mask: npt.ArrayLike,
     baseline_lengths: npt.ArrayLike,  # Number of samples per baseline
 ) -> None:
     """
@@ -533,12 +555,20 @@ def _update_sum_map_with_tod(
     """
 
     for det_idx in range(pixel_idx.shape[0]):
+        if not d_mask[det_idx]:
+            continue
         cur_weight = weights[det_idx]
 
         baseline_idx = 0
         samples_in_this_baseline = 0
 
         for sample_idx in range(tod.shape[1]):
+            if not t_mask[sample_idx]:
+                (baseline_idx, samples_in_this_baseline) = _step_over_baseline(
+                    baseline_idx, samples_in_this_baseline, baseline_lengths
+                )
+                continue
+
             cur_pix = pixel_idx[det_idx, sample_idx]
             _sum_map_contribution_from_one_sample(
                 pol_angle_rad=pol_angle_rad[det_idx, sample_idx],
@@ -560,6 +590,8 @@ def _update_sum_map_with_baseline(
     pol_angle_rad: npt.ArrayLike,
     pixel_idx: npt.ArrayLike,
     weights: npt.ArrayLike,
+    d_mask: npt.ArrayLike,
+    t_mask: npt.ArrayLike,
     baselines: npt.ArrayLike,  # Value of each baseline
     baseline_lengths: npt.ArrayLike,  # Number of samples per baseline
 ) -> None:
@@ -574,12 +606,20 @@ def _update_sum_map_with_baseline(
     """
 
     for det_idx in range(pixel_idx.shape[0]):
+        if not d_mask[det_idx]:
+            continue
         cur_weight = weights[det_idx]
 
         baseline_idx = 0
         samples_in_this_baseline = 0
 
         for sample_idx in range(pixel_idx.shape[1]):
+            if not t_mask[sample_idx]:
+                (baseline_idx, samples_in_this_baseline) = _step_over_baseline(
+                    baseline_idx, samples_in_this_baseline, baseline_lengths
+                )
+                continue
+
             cur_pix = pixel_idx[det_idx, sample_idx]
             _sum_map_contribution_from_one_sample(
                 pol_angle_rad=pol_angle_rad[det_idx, sample_idx],
@@ -625,6 +665,8 @@ def _compute_binned_map(
     nobs_matrix_cholesky: NobsMatrix,
     baselines_list: Optional[List[npt.ArrayLike]],
     baseline_lengths_list: List[npt.ArrayLike],
+    dm_list: List[npt.ArrayLike],
+    tm_list: List[npt.ArrayLike],
     component: Optional[str],
     output_hit_map: npt.ArrayLike,
     output_sky_map: npt.ArrayLike,
@@ -661,8 +703,8 @@ def _compute_binned_map(
     output_sky_map[:] = 0
     output_hit_map[:] = 0
 
-    for obs_idx, (cur_obs, cur_baseline_lengths) in enumerate(
-        zip(obs_list, baseline_lengths_list)
+    for obs_idx, (cur_obs, cur_baseline_lengths, cur_d_mask, cur_t_mask) in enumerate(
+        zip(obs_list, baseline_lengths_list, dm_list, tm_list)
     ):
         if baselines_list is None:
             _update_sum_map_with_tod(
@@ -672,6 +714,8 @@ def _compute_binned_map(
                 pol_angle_rad=cur_obs.destriper_pol_angle_rad,
                 pixel_idx=cur_obs.destriper_pixel_idx,
                 weights=cur_obs.destriper_weights,
+                d_mask=cur_d_mask,
+                t_mask=cur_t_mask,
                 baseline_lengths=cur_baseline_lengths,
             )
         else:
@@ -682,6 +726,8 @@ def _compute_binned_map(
                 pol_angle_rad=cur_obs.destriper_pol_angle_rad,
                 pixel_idx=cur_obs.destriper_pixel_idx,
                 weights=cur_obs.destriper_weights,
+                d_mask=cur_d_mask,
+                t_mask=cur_t_mask,
                 baselines=cur_baselines,
                 baseline_lengths=cur_baseline_lengths,
             )
@@ -716,6 +762,8 @@ def _compute_tod_sums_for_one_component(
     pixel_idx: npt.ArrayLike,
     psi_angle_rad: npt.ArrayLike,
     sky_map: npt.ArrayLike,
+    d_mask: npt.ArrayLike,
+    t_mask: npt.ArrayLike,
     baseline_length: npt.ArrayLike,
     output_sums: npt.ArrayLike,
 ) -> None:
@@ -734,6 +782,8 @@ def _compute_tod_sums_for_one_component(
     output_sums[:] = 0
 
     for det_idx, cur_weight in enumerate(weights):
+        if not d_mask[det_idx]:
+            continue
         det_pixel_idx = pixel_idx[det_idx, :]
         det_psi_angle_rad = psi_angle_rad[det_idx, :]
 
@@ -741,6 +791,12 @@ def _compute_tod_sums_for_one_component(
         samples_in_this_baseline = 0
 
         for sample_idx in range(len(det_pixel_idx)):
+            if not t_mask[sample_idx]:
+                (baseline_idx, samples_in_this_baseline) = _step_over_baseline(
+                    baseline_idx, samples_in_this_baseline, baseline_length
+                )
+                continue
+
             map_value = estimate_sample_from_map(
                 cur_pixel=det_pixel_idx[sample_idx],
                 cur_psi=det_psi_angle_rad[sample_idx],
@@ -761,6 +817,8 @@ def _compute_baseline_sums_for_one_component(
     pixel_idx: npt.ArrayLike,
     psi_angle_rad: npt.ArrayLike,
     sky_map: npt.ArrayLike,
+    d_mask: npt.ArrayLike,
+    t_mask: npt.ArrayLike,
     baselines: npt.ArrayLike,
     baseline_length: npt.ArrayLike,
     output_sums: npt.ArrayLike,
@@ -780,6 +838,8 @@ def _compute_baseline_sums_for_one_component(
     output_sums[:] = 0
 
     for det_idx, cur_weight in enumerate(weights):
+        if not d_mask[det_idx]:
+            continue
         det_pixel_idx = pixel_idx[det_idx, :]
         det_psi_angle_rad = psi_angle_rad[det_idx, :]
 
@@ -787,6 +847,12 @@ def _compute_baseline_sums_for_one_component(
         samples_in_this_baseline = 0
 
         for sample_idx in range(len(det_pixel_idx)):
+            if not t_mask[sample_idx]:
+                (baseline_idx, samples_in_this_baseline) = _step_over_baseline(
+                    baseline_idx, samples_in_this_baseline, baseline_length
+                )
+                continue
+
             map_value = estimate_sample_from_map(
                 cur_pixel=det_pixel_idx[sample_idx],
                 cur_psi=det_psi_angle_rad[sample_idx],
@@ -807,6 +873,8 @@ def _compute_baseline_sums(
     baselines_list: Optional[List[npt.ArrayLike]],
     baseline_lengths_list: List[npt.ArrayLike],
     component: Optional[str],
+    dm_list: List[npt.ArrayLike],
+    tm_list: List[npt.ArrayLike],
     output_sums_list: List[npt.ArrayLike],
 ):
     """
@@ -842,8 +910,14 @@ def _compute_baseline_sums(
     )
 
     # Compute the value of the F^t C_w⁻¹ Z operator
-    for obs_idx, (cur_obs, cur_baseline_lengths, cur_sums) in enumerate(
-        zip(obs_list, baseline_lengths_list, output_sums_list)
+    for obs_idx, (
+        cur_obs,
+        cur_baseline_lengths,
+        cur_sums,
+        cur_d_mask,
+        cur_t_mask,
+    ) in enumerate(
+        zip(obs_list, baseline_lengths_list, output_sums_list, dm_list, tm_list)
     ):
         assert len(cur_baseline_lengths) == cur_sums.shape[1], (
             f"The output buffer for observation {obs_idx=} "
@@ -863,6 +937,8 @@ def _compute_baseline_sums(
                 pixel_idx=cur_obs.destriper_pixel_idx,
                 psi_angle_rad=cur_obs.destriper_pol_angle_rad,
                 sky_map=sky_map,
+                d_mask=cur_d_mask,
+                t_mask=cur_t_mask,
                 baselines=cur_baseline,
                 baseline_length=cur_baseline_lengths,
                 output_sums=cur_sums,
@@ -874,6 +950,8 @@ def _compute_baseline_sums(
                 pixel_idx=cur_obs.destriper_pixel_idx,
                 psi_angle_rad=cur_obs.destriper_pol_angle_rad,
                 sky_map=sky_map,
+                d_mask=cur_d_mask,
+                t_mask=cur_t_mask,
                 baseline_length=cur_baseline_lengths,
                 output_sums=cur_sums,
             )
@@ -923,6 +1001,8 @@ def _compute_b_or_Ax(
     nobs_matrix_cholesky: NobsMatrix,
     sky_map: npt.ArrayLike,
     hit_map: npt.ArrayLike,
+    dm_list: List[npt.ArrayLike],
+    tm_list: List[npt.ArrayLike],
     baselines_list: Optional[List[npt.ArrayLike]],
     baseline_lengths_list: List[npt.ArrayLike],
     component: Optional[str],
@@ -940,6 +1020,8 @@ def _compute_b_or_Ax(
         nobs_matrix_cholesky=nobs_matrix_cholesky,
         baselines_list=baselines_list,
         baseline_lengths_list=baseline_lengths_list,
+        dm_list=dm_list,
+        tm_list=tm_list,
         component=component,
         output_sky_map=sky_map,
         output_hit_map=hit_map,
@@ -951,6 +1033,8 @@ def _compute_b_or_Ax(
         baselines_list=baselines_list,
         baseline_lengths_list=baseline_lengths_list,
         component=component,
+        dm_list=dm_list,
+        tm_list=tm_list,
         output_sums_list=result,
     )
 
@@ -962,6 +1046,8 @@ def compute_b(
     hit_map: npt.ArrayLike,
     baseline_lengths_list: List[npt.ArrayLike],
     component: str,
+    dm_list: List[npt.ArrayLike],
+    tm_list: List[npt.ArrayLike],
     result: List[npt.ArrayLike],
 ) -> None:
     """
@@ -976,6 +1062,8 @@ def compute_b(
         nobs_matrix_cholesky=nobs_matrix_cholesky,
         sky_map=sky_map,
         hit_map=hit_map,
+        dm_list=dm_list,
+        tm_list=tm_list,
         baselines_list=None,
         baseline_lengths_list=baseline_lengths_list,
         component=component,
@@ -988,6 +1076,8 @@ def compute_Ax(
     nobs_matrix_cholesky: NobsMatrix,
     sky_map: npt.ArrayLike,
     hit_map: npt.ArrayLike,
+    dm_list: List[npt.ArrayLike],
+    tm_list: List[npt.ArrayLike],
     baselines_list: List[npt.ArrayLike],
     baseline_lengths_list: List[npt.ArrayLike],
     result: List[npt.ArrayLike],
@@ -1005,6 +1095,8 @@ def compute_Ax(
         nobs_matrix_cholesky=nobs_matrix_cholesky,
         sky_map=sky_map,
         hit_map=hit_map,
+        dm_list=dm_list,
+        tm_list=tm_list,
         baselines_list=baselines_list,
         baseline_lengths_list=baseline_lengths_list,
         component=None,
@@ -1015,7 +1107,7 @@ def compute_Ax(
 @njit
 def compute_weighted_baseline_length(
     lengths: npt.ArrayLike, weights: npt.ArrayLike, result: npt.ArrayLike
-) -> float:
+) -> None:
     """Compute Σ Nᵢ/σᵢ, where the summation is done over the detectors
 
     This quantity is used both to estimate the error bar for each baseline
@@ -1023,13 +1115,13 @@ def compute_weighted_baseline_length(
     algorithm.
     """
 
-    for det_idx in range(len(result)):
-        result[det_idx] = 0.0
+    for baseline_idx in range(len(result)):
+        result[baseline_idx] = 0.0
         for cur_weight in weights:
-            result[det_idx] += lengths[det_idx] / cur_weight
+            result[baseline_idx] += lengths[baseline_idx] / cur_weight
 
         # We store the *inverse*, i.e., the diagonal of matrix M⁻¹
-        result[det_idx] = 1.0 / result[det_idx]
+        result[baseline_idx] = 1.0 / result[baseline_idx]
 
 
 def _create_preconditioner(obs_list, baseline_lengths_list) -> List[npt.ArrayLike]:
@@ -1081,12 +1173,16 @@ def _run_destriper(
     hit_map: npt.ArrayLike,
     baseline_lengths_list: List[npt.ArrayLike],
     baselines_list_start: List[npt.ArrayLike],
+    dm_list: List[npt.ArrayLike],
+    tm_list: List[npt.ArrayLike],
     component: str,
     threshold: float,
     max_steps: int,
     use_preconditioner: bool,
     callback: Any,
     callback_kwargs: Dict[Any, Any],
+    recycled_convergence: bool,
+    recycle_baselines: Optional[bool] = False,
 ) -> Tuple[
     List[npt.ArrayLike],  # The solution, i.e., the list of baselines
     List[npt.ArrayLike],  # The error bars of the baselines
@@ -1166,6 +1262,8 @@ def _run_destriper(
         nobs_matrix_cholesky=nobs_matrix_cholesky,
         sky_map=destriped_map,
         hit_map=hit_map,
+        dm_list=dm_list,
+        tm_list=tm_list,
         baseline_lengths_list=baseline_lengths_list,
         component=component,
         result=b,
@@ -1175,6 +1273,8 @@ def _run_destriper(
         nobs_matrix_cholesky=nobs_matrix_cholesky,
         sky_map=destriped_map,
         hit_map=hit_map,
+        dm_list=dm_list,
+        tm_list=tm_list,
         baselines_list=x,
         baseline_lengths_list=baseline_lengths_list,
         result=Ax,
@@ -1213,12 +1313,17 @@ def _run_destriper(
         if k >= max_steps:
             converged = False
             break
+        if recycle_baselines:
+            converged = f"Recycled baselines with 'converged = {recycled_convergence}' were used!"
+            break
 
         compute_Ax(
             obs_list=obs_list,
             nobs_matrix_cholesky=nobs_matrix_cholesky,
             sky_map=destriped_map,
             hit_map=hit_map,
+            dm_list=dm_list,
+            tm_list=tm_list,
             baselines_list=z,
             baseline_lengths_list=baseline_lengths_list,
             result=Ax,
@@ -1271,6 +1376,8 @@ def _run_destriper(
         output_hit_map=hit_map,
         nobs_matrix_cholesky=nobs_matrix_cholesky,
         component=component,
+        dm_list=dm_list,
+        tm_list=tm_list,
         baselines_list=None,
         baseline_lengths_list=baseline_lengths_list,
     )
@@ -1281,6 +1388,8 @@ def _run_destriper(
         output_sky_map=destriped_map,
         output_hit_map=hit_map,
         nobs_matrix_cholesky=nobs_matrix_cholesky,
+        dm_list=dm_list,
+        tm_list=tm_list,
         component=None,
         baselines_list=best_x,
         baseline_lengths_list=baseline_lengths_list,
@@ -1335,6 +1444,10 @@ def make_destriped_map(
     keep_weights: bool = False,
     keep_pixel_idx: bool = False,
     keep_pol_angle_rad: bool = False,
+    detector_split: str = "full",
+    time_split: str = "full",
+    baselines_list: Optional[List[npt.ArrayLike]] = None,
+    recycled_convergence: bool = False,
     callback: Any = destriper_log_callback,
     callback_kwargs: Optional[Dict[Any, Any]] = None,
 ) -> DestriperResult:
@@ -1471,11 +1584,17 @@ def make_destriped_map(
             factor=+1.0,
         )
 
+    detector_mask_list = _build_mask_detector_split(detector_split, obs_list)
+
+    time_mask_list = _build_mask_time_split(time_split, obs_list)
+
     nobs_matrix_cholesky = _build_nobs_matrix(
         hpx=hpx,
         obs_list=obs_list,
         ptg_list=ptg_list,
         psi_list=psi_list,
+        dm_list=detector_mask_list,
+        tm_list=time_mask_list,
     )
 
     number_of_pixels = hpx.npix()
@@ -1506,10 +1625,14 @@ def make_destriped_map(
 
         # Each element of this list is a 2D array with shape (N_det, N_baselines),
         # where N_det is the number of detectors in the i-th Observation object
-        baselines_list = [
-            np.zeros((getattr(cur_obs, components[0]).shape[0], len(cur_baseline)))
-            for (cur_obs, cur_baseline) in zip(obs_list, baseline_lengths_list)
-        ]
+        recycle_baselines = False
+        if baselines_list is None:
+            baselines_list = [
+                np.zeros((getattr(cur_obs, components[0]).shape[0], len(cur_baseline)))
+                for (cur_obs, cur_baseline) in zip(obs_list, baseline_lengths_list)
+            ]
+        else:
+            recycle_baselines = True
 
         destriped_map = np.empty((3, number_of_pixels))
         (
@@ -1527,6 +1650,10 @@ def make_destriped_map(
             hit_map=hit_map,
             baseline_lengths_list=baseline_lengths_list,
             baselines_list_start=baselines_list,
+            recycle_baselines=recycle_baselines,
+            recycled_convergence=recycled_convergence,
+            dm_list=detector_mask_list,
+            tm_list=time_mask_list,
             component=components[0],
             threshold=params.threshold,
             max_steps=params.iter_max,
@@ -1549,6 +1676,8 @@ def make_destriped_map(
             output_hit_map=hit_map,
             nobs_matrix_cholesky=nobs_matrix_cholesky,
             component=components[0],
+            dm_list=detector_mask_list,
+            tm_list=time_mask_list,
             baselines_list=None,
             baseline_lengths_list=[
                 np.array([getattr(cur_obs, components[0]).shape[1]], dtype=int)
@@ -1608,6 +1737,8 @@ def make_destriped_map(
         binned_map=binned_map,
         nobs_matrix_cholesky=nobs_matrix_cholesky,
         coordinate_system=params.output_coordinate_system,
+        detector_split=detector_split,
+        time_split=time_split,
         # The following fields are filled only if the CG algorithm was used
         baselines=baselines_list,
         baseline_errors=baseline_errors_list,
@@ -1731,6 +1862,10 @@ def _save_rank0_destriper_results(results: DestriperResult, output_file: Path) -
         "Size of temporary buffers [bytes]",
     )
     primary_hdu.header["ELAPSEDT"] = (results.elapsed_time_s, "Wall clock time [s]")
+
+    primary_hdu.header["DSPLIT"] = (results.detector_split, "Detector split")
+
+    primary_hdu.header["TSPLIT"] = (results.detector_split, "Time split")
 
     hdu_list = [primary_hdu]
 
@@ -1875,15 +2010,13 @@ def _save_baselines(results: DestriperResult, output_file: Path) -> None:
         error_hdu = fits.BinTableHDU.from_columns(
             [
                 fits.Column(
-                    name=f"ERR{det_idx:05d}",
-                    array=cur_error[det_idx, :],
+                    name=f"ERR{idx:05d}",
+                    array=cur_error,
                     format="1E",
                     unit="K",
-                )
-                for det_idx in range(cur_baseline.shape[0])
+                ),
             ]
         )
-        error_hdu.header["NUMDETS"] = (cur_error.shape[0], "Number of detectors")
 
         length_hdu = fits.BinTableHDU.from_columns(
             [fits.Column(name="LENGTH", array=cur_lengths, unit="", format="1J")]
@@ -1903,7 +2036,6 @@ def _save_baselines(results: DestriperResult, output_file: Path) -> None:
                 "Number of samples covered by these baselines",
             )
             this_hdu.name = f"{hdu_base_name}{idx:05d}"
-            print(f"{this_hdu.name=}")
 
         hdu_list += [baseline_hdu, error_hdu, length_hdu]
         idx += 1
@@ -1914,7 +2046,12 @@ def _save_baselines(results: DestriperResult, output_file: Path) -> None:
         fits.HDUList(hdu_list).writeto(outf, overwrite=True)
 
 
-def save_destriper_results(results: DestriperResult, output_folder: Path) -> None:
+def save_destriper_results(
+    results: DestriperResult,
+    output_folder: Path,
+    custom_dest_file: Optional[str] = None,
+    custom_base_file: Optional[str] = None,
+) -> None:
     """
     Save the results of a call to :func:`.make_destriped_map` to disk
 
@@ -1941,13 +2078,22 @@ def save_destriper_results(results: DestriperResult, output_folder: Path) -> Non
 
     # Only MPI process #0 saves the file with the maps
     if MPI_COMM_WORLD.rank == 0:
-        _save_rank0_destriper_results(
-            results=results, output_file=output_folder / __DESTRIPER_RESULTS_FILE_NAME
-        )
+        if custom_dest_file:
+            _save_rank0_destriper_results(
+                results=results, output_file=output_folder / custom_dest_file
+            )
+        else:
+            _save_rank0_destriper_results(
+                results=results,
+                output_file=output_folder / __DESTRIPER_RESULTS_FILE_NAME,
+            )
 
     # Now let's save the baselines: one per each observation
     if results.destriped_map is not None:
-        _save_baselines(results, output_file=output_folder / __BASELINES_FILE_NAME)
+        if custom_base_file:
+            _save_baselines(results, output_file=output_folder / custom_base_file)
+        else:
+            _save_baselines(results, output_file=output_folder / __BASELINES_FILE_NAME)
 
 
 def _load_rank0_destriper_results(file_path: Path) -> DestriperResult:
@@ -1982,6 +2128,8 @@ def _load_rank0_destriper_results(file_path: Path) -> DestriperResult:
                 [inpf["BINMAP"].data.field(comp) for comp in ("I", "Q", "U")]
             ),
             coordinate_system=coord_sys,
+            detector_split=inpf[0].header["DSPLIT"],
+            time_split=inpf[0].header["TSPLIT"],
             history_of_stopping_factors=[],
             elapsed_time_s=inpf[0].header["ELAPSEDT"],
             destriped_map=None,
@@ -2012,7 +2160,11 @@ def _load_rank0_destriper_results(file_path: Path) -> DestriperResult:
     return result
 
 
-def load_destriper_results(folder: Path) -> DestriperResult:
+def load_destriper_results(
+    folder: Path,
+    custom_dest_file: Optional[str] = None,
+    custom_base_file: Optional[str] = None,
+) -> DestriperResult:
     """
     Load the results of a call to :func:`.make_destriped_map` from disk
 
@@ -2032,14 +2184,20 @@ def load_destriper_results(folder: Path) -> DestriperResult:
 
     # We run this on *all* the MPI processes, as it might be that each of them
     # needs this information!
-    result = _load_rank0_destriper_results(folder / __DESTRIPER_RESULTS_FILE_NAME)
+    if custom_dest_file:
+        result = _load_rank0_destriper_results(folder / custom_dest_file)
+    else:
+        result = _load_rank0_destriper_results(folder / __DESTRIPER_RESULTS_FILE_NAME)
 
     if result.destriped_map is not None:
         result.baselines = []
         result.baseline_errors = []
         result.baseline_lengths = []
 
-        baselines_file_name = folder / __BASELINES_FILE_NAME
+        if custom_base_file:
+            baselines_file_name = folder / custom_base_file
+        else:
+            baselines_file_name = folder / __BASELINES_FILE_NAME
 
         with fits.open(baselines_file_name) as inpf:
             assert MPI_COMM_WORLD.rank == inpf[0].header["MPIRANK"], (
@@ -2064,12 +2222,7 @@ def load_destriper_results(folder: Path) -> DestriperResult:
                     ),
                 )
                 result.baseline_errors.append(
-                    np.array(
-                        [
-                            inpf[f"ERR{obs_idx:05d}"].data.field(f"ERR{det_idx:05d}")
-                            for det_idx in range(num_of_detectors)
-                        ]
-                    ),
+                    np.array(inpf[f"ERR{obs_idx:05d}"].data.field(f"ERR{obs_idx:05d}")),
                 )
                 result.baseline_lengths.append(
                     inpf[f"LEN{obs_idx:05d}"].data.field("LENGTH")