+++ test/test_destriper.py
@@ -372,6 +372,8 @@ def test_map_maker_parts():
     from litebird_sim.mapmaking.common import _normalize_observations_and_pointings
     from litebird_sim.mapmaking.destriper import (
         _store_pixel_idx_and_pol_angle_in_obs,
+        _build_mask_detector_split,
+        _build_mask_time_split,
         _build_nobs_matrix,
         _sum_components_into_obs,
         _compute_binned_map,
@@ -401,6 +403,10 @@ def test_map_maker_parts():
         output_coordinate_system=CoordinateSystem.Ecliptic,
     )
 
+    detector_mask_list = _build_mask_detector_split("full", obs_list)
+
+    time_mask_list = _build_mask_time_split("full", obs_list)
+
     #################################################
     # Step 1: check that the N_obs matrix is correct
 
@@ -409,6 +415,8 @@ def test_map_maker_parts():
         obs_list=obs_list,
         ptg_list=ptg_list,
         psi_list=psi_list,
+        dm_list=detector_mask_list,
+        tm_list=time_mask_list,
     )
 
     assert (
@@ -460,6 +468,8 @@ def test_map_maker_parts():
         baselines_list=None,
         baseline_lengths_list=baseline_lengths_list,
         component="sky_signal",
+        dm_list=detector_mask_list,
+        tm_list=time_mask_list,
     )
 
     # This is going to be a 3N_p vector
@@ -510,6 +520,8 @@ def test_map_maker_parts():
         baselines_list=baselines_list,
         baseline_lengths_list=baseline_lengths_list,
         component=None,
+        dm_list=detector_mask_list,
+        tm_list=time_mask_list,
     )
 
     _compute_baseline_sums(
@@ -519,6 +531,8 @@ def test_map_maker_parts():
         baseline_lengths_list=baseline_lengths_list,
         component="sky_signal",
         output_sums_list=output_baselines_list,
+        dm_list=detector_mask_list,
+        tm_list=time_mask_list,
     )
 
     expected = (
@@ -549,6 +563,8 @@ def test_map_maker_parts():
         baselines_list=None,
         baseline_lengths_list=baseline_lengths_list,
         component="sky_signal",
+        dm_list=detector_mask_list,
+        tm_list=time_mask_list,
     )
 
     _compute_baseline_sums(
@@ -558,6 +574,8 @@ def test_map_maker_parts():
         baseline_lengths_list=baseline_lengths_list,
         component="sky_signal",
         output_sums_list=output_baselines_list,
+        dm_list=detector_mask_list,
+        tm_list=time_mask_list,
     )
 
     expected = (
@@ -864,7 +882,7 @@ def _test_destriper_results_io(tmp_path, use_destriper: bool):
     if use_destriper:
         baselines = [np.random.random((num_of_detectors, num_of_baselines))]
         baselines[0] -= np.mean(baselines[0])  # Make their mean zero
-        baseline_errors = [1.5 + np.random.random((num_of_detectors, num_of_baselines))]
+        baseline_errors = [1.5 + np.random.random((num_of_baselines))]
         baseline_lengths = [150 + np.arange(num_of_baselines, dtype=int) * 4]
         iter_max = 123
         threshold = 1.2345e-6
@@ -907,6 +925,8 @@ def _test_destriper_results_io(tmp_path, use_destriper: bool):
         binned_map=binned_map,
         nobs_matrix_cholesky=nobs_matrix_cholesky,
         coordinate_system=lbs.CoordinateSystem.Galactic,
+        detector_split="full",
+        time_split="full",
         baselines=baselines,
         baseline_errors=baseline_errors,
         baseline_lengths=baseline_lengths,
@@ -1011,3 +1031,88 @@ def test_destriper_io(tmp_path):
 
 def test_destriper_io_without_destriper(tmp_path):
     _test_destriper_results_io(tmp_path=tmp_path, use_destriper=False)
+
+
+def test_save_baselines_for_many_detectors(tmp_path):
+    nside = 32
+
+    sim = lbs.Simulation(
+        base_path=Path(tmp_path) / "test_errors_dimension",
+        start_time=0,
+        duration_s=astropy.time.TimeDelta(1, format="jd").to("s").value,
+        random_seed=12345,
+    )
+
+    sim.set_instrument(
+        lbs.InstrumentInfo(
+            name="Dummy", boresight_rotangle_rad=np.deg2rad(50), hwp_rpm=46.0
+        )
+    )
+
+    dets = [
+        lbs.DetectorInfo(
+            sampling_rate_hz=1.0,
+            name="A",
+            fwhm_arcmin=20.0,
+            bandcenter_ghz=140.0,
+            bandwidth_ghz=40.0,
+            net_ukrts=50.0,
+            fknee_mhz=50.0,
+            quat=np.array([0.02568196, 0.00506653, 0.0, 0.99965732]),
+        ),
+        lbs.DetectorInfo(
+            sampling_rate_hz=1.0,
+            name="B",
+            fwhm_arcmin=20.0,
+            bandcenter_ghz=140.0,
+            bandwidth_ghz=40.0,
+            net_ukrts=50.0,
+            fknee_mhz=50.0,
+            quat=np.array([0.0145773, 0.02174247, -0.70686447, 0.70686447]),
+        ),
+    ]
+
+    sim.set_scanning_strategy(
+        scanning_strategy=lbs.SpinningScanningStrategy(
+            spin_sun_angle_rad=np.deg2rad(45.0),
+            precession_rate_hz=1 / 10_020.0,
+            spin_rate_hz=1 / 60.0,
+        ),
+    )
+
+    sim.create_observations(
+        detectors=dets,
+        num_of_obs_per_detector=sim.mpi_comm.size,
+    )
+
+    assert len(sim.observations) == 1
+
+    sim.set_hwp(
+        lbs.IdealHWP(
+            sim.instrument.hwp_rpm * 2 * np.pi / 60,
+        ),  # applies hwp rotation angle to the polarization angle
+    )
+    sim.compute_pointings()
+
+    lbs.add_noise_to_observations(
+        obs=sim.observations,
+        noise_type="one_over_f",
+        scale=1,
+        random=sim.random,
+    )
+
+    destriper_params_noise = lbs.DestriperParameters(
+        output_coordinate_system=lbs.coordinates.CoordinateSystem.Galactic,
+        samples_per_baseline=100,  # ν_samp = 1 Hz ⇒ the baseline is 100 s
+        iter_max=10,
+        threshold=1e-6,
+    )
+
+    destriper_result = lbs.make_destriped_map(
+        nside=nside, obs=sim.observations, pointings=None, params=destriper_params_noise
+    )
+
+    save_destriper_results(
+        output_folder=Path(tmp_path) / "test_errors_dimension",
+        results=destriper_result,
+    )